# api/docs/swagger.yaml
definitions:
...
  models.BookCreate:
    properties:
      book_file:
        type: string
      level_id:
        type: integer
      name:
        type: string
      picture:
        type: string
    type: object
  models.BookResponse:
    properties:
      book_file:
        type: string
      created_at:
        type: string
      id:
        type: integer
      lesson_name:
        type: string
      level_id:
        type: integer
      picture:
        type: string
      updated_at:
        type: string
    type: object
  models.BookUpdate:
    properties:
      book_file:
        type: string
      id:
        type: integer
      level_id:
        type: integer
      name:
        type: string
      picture:
        type: string
    type: object
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  title: Asrlan-Monolithic
paths:
...
  /v1/book:
    post:
      consumes:
      - application/json
      description: This Api for creating a new book
      parameters:
      - description: BookCreate Model
        in: body
        name: BookCreate
        required: true
        schema:
          $ref: '#/definitions/models.BookCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Create Book
      tags:
      - books
    put:
      consumes:
      - application/json
      description: This Api for updating book
      parameters:
      - description: Update BookUpdate Model
        in: body
        name: BookUpdate
        required: true
        schema:
          $ref: '#/definitions/models.BookUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Update Book
      tags:
      - books
  /v1/book/{id}:
    delete:
      consumes:
      - application/json
      description: This Api for deleting book
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete Book
      tags:
      - books
    get:
      consumes:
      - application/json
      description: This Api for get book
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Get Book
      tags:
      - books
  /v1/books/{id}:
    get:
      consumes:
      - application/json
      description: This Api for get all books
      parameters:
      - description: Level Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: ListBooks
      tags:
      - books
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"